feat: Integrate redux on Cart Empty page
